apps/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv
sqlalchemy>=2.0
structlog>=23.2.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.21.0
